# docs/marshaller_swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Marshaller API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/path-options": {
            "get": {
                "description": "returns the matching paths per device-type. Naming more than one aspect is an AND on one content variable, and each aspect covers its own subtree.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "path-options"
                ],
                "summary": "list the paths of device-types that match a function and aspects",
                "parameters": [
                    {
                        "type": "string",
                        "description": "comma separated device-type ids; an absent parameter yields an empty result",
                        "name": "device-type-ids",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "comma separated characteristic ids; an absent parameter yields an empty result",
                        "name": "characteristic-filter",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the function the path has to serve",
                        "name": "function-id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "comma separated aspect ids",
                        "name": "aspect-ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "deprecated: use aspect-ids; an alias for a list with one element",
                        "name": "aspect-id",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "omit the value. prefix from the returned paths",
                        "name": "without-envelope",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/marshaller.PathOptionsResultElement"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "without-envelope is not a bool",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/query/path-options": {
            "post": {
                "description": "the request-body form of GET /path-options. An empty characteristic_id_filter means \"do not filter\" here, unlike the absent query parameter.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "path-options"
                ],
                "summary": "list the paths of device-types that match a function and aspects",
                "parameters": [
                    {
                        "description": "the device-types, function and aspects to match",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.PathOptionsQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/marshaller.PathOptionsResultElement"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "the body is not valid json",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/unmarshal": {
            "post": {
                "description": "reads the message of the given service and returns its value in the requested characteristic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal"
                ],
                "summary": "unmarshal a protocol message into a value",
                "parameters": [
                    {
                        "description": "service, protocol and the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/unmarshal/{serviceId}/{characteristicId}": {
            "post": {
                "description": "like POST /unmarshal, but the service is read from the device-repository by the id in the path and the characteristic comes from the path as well",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal"
                ],
                "summary": "unmarshal a protocol message, service and characteristic from the path",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of the service the message came from",
                        "name": "serviceId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id of the characteristic the value is wanted in",
                        "name": "characteristicId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, or the service does not reference the given protocol",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/v2/unmarshal": {
            "post": {
                "description": "returns the value at the requested path; without a path it is determined from the function and the requested aspects, closest aspect first",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "unmarshal",
                    "v2"
                ],
                "summary": "unmarshal a protocol message into a value",
                "parameters": [
                    {
                        "description": "service, protocol and the message to unmarshal",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/messages.UnmarshallingV2Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the value in the requested characteristic",
                        "schema": {}
                    },
                    "400": {
                        "description": "the body is not valid json, the service does not reference the given protocol, or no output path matches the criteria",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "marshaller.PathOptionsResultElement": {
            "type": "object",
            "properties": {
                "json_path": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "path_to_characteristic_id": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "service_id": {
                    "type": "string"
                }
            }
        },
...
        "messages.PathOptionsQuery": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "description": "deprecated: please use AspectIds",
                    "type": "string"
                },
                "aspect_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "characteristic_id_filter": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "device_type_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "function_id": {
                    "type": "string"
                },
                "without_envelope": {
                    "type": "boolean"
                }
            }
        },
        "messages.UnmarshallingRequest": {
            "type": "object",
            "properties": {
                "characteristic_id": {
                    "description": "semi-optional, may be determined by request path",
                    "type": "string"
                },
                "content_variable_hints": {
                    "description": "optional",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path_allow_list": {
                    "description": "optional\n\t\tif len \u003e 0: apply data only on given ContentVariable paths\n\t\tuseful if 2 variables have the same characteristic assigned but are used with different functions or aspects",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "protocol": {
                    "description": "semi-optional, may be determined by service",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
        "messages.UnmarshallingV2Request": {
            "type": "object",
            "properties": {
                "aspect_node": {
                    "description": "deprecated: please use AspectNodes",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.AspectNode"
                        }
                    ]
                },
                "aspect_node_id": {
                    "description": "deprecated: please use AspectNodeIds",
                    "type": "string"
                },
                "aspect_node_ids": {
                    "description": "semi-optional, to determine AspectNodes if not set",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aspect_nodes": {
                    "description": "semi-optional, to determine Path if not set, may themselves be determent by AspectNodeIds",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.AspectNode"
                    }
                },
                "characteristic_id": {
                    "description": "semi-optional, may be determined by request path",
                    "type": "string"
                },
                "function_id": {
                    "description": "semi-optional, to determine Path if not set",
                    "type": "string"
                },
                "message": {
                    "description": "semi-optional; may be needed to create serialized_output",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "path": {
                    "description": "semi-optional, may be determent by FunctionId and AspectNodes",
                    "type": "string"
                },
                "protocol": {
                    "description": "semi-optional, may be determined by service",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Protocol"
                        }
                    ]
                },
                "serialized_output": {
                    "description": "semi-optional; may be created from message",
                    "type": "object",
                    "additionalProperties": true
                },
                "service": {
                    "description": "semi-optional, may be determined by request path",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Service"
                        }
                    ]
                }
            }
        },
        "model.AspectNode": {
            "type": "object",
            "properties": {
                "ancestor_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "aspect_class_id": {
                    "type": "string"
                },
                "child_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "descendent_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_id": {
                    "type": "string"
                },
                "root_id": {
                    "type": "string"
                }
            }
        },
...
        "model.Protocol": {
            "type": "object",
            "properties": {
                "constraints": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "handler": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "protocol_segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProtocolSegment"
                    }
                }
            }
        },
        "model.Service": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Attribute"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "inputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "interaction": {
                    "$ref": "#/definitions/models.Interaction"
                },
                "local_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "outputs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Content"
                    }
                },
                "protocol_id": {
                    "type": "string"
                },
                "service_group_key": {
                    "type": "string"
                }
            }
        },
        "models.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "origin": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.Content": {
            "type": "object",
            "properties": {
                "content_variable": {
                    "$ref": "#/definitions/models.ContentVariable"
                },
                "id": {
                    "type": "string"
                },
                "protocol_segment_id": {
                    "type": "string"
                },
                "serialization": {
                    "$ref": "#/definitions/models.Serialization"
                }
            }
        },
        "models.ContentVariable": {
            "type": "object",
            "properties": {
                "aspect_id": {
                    "description": "deprecated: please use AspectIds",
                    "type": "string"
                },
                "aspect_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "characteristic_id": {
                    "type": "string"
                },
                "function_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_void": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "omit_empty": {
                    "type": "boolean"
                },
                "serialization_options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sub_content_variables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContentVariable"
                    }
                },
                "type": {
                    "$ref": "#/definitions/models.Type"
                },
                "unit_reference": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "models.Interaction": {
            "type": "string",
            "enum": [
                "event",
                "request",
                "event+request"
            ],
            "x-enum-varnames": [
                "EVENT",
                "REQUEST",
                "EVENT_AND_REQUEST"
            ]
        },
        "models.ProtocolSegment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Serialization": {
            "type": "string",
            "enum": [
                "xml",
                "json",
                "plain-text"
            ],
            "x-enum-varnames": [
                "XML",
                "JSON",
                "PlainText"
            ]
        },
        "models.Type": {
            "type": "string",
            "enum": [
                "https://schema.org/Text",
                "https://schema.org/Integer",
                "https://schema.org/Float",
                "https://schema.org/Boolean",
                "https://schema.org/ItemList",
                "https://schema.org/StructuredValue"
            ],
            "x-enum-varnames": [
                "String",
                "Integer",
                "Float",
                "Boolean",
                "List",
                "Structure"
            ]
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}